Sửa UI, mã hóa tk mk

+++ AssemblyCSharp/Rms.cs
@@ -1,5 +1,7 @@
 using System;
 using System.IO;
+using System.Security.Cryptography;
+using System.Text;
 using System.Threading;
 using UnityEngine;
 
@@ -36,27 +38,6 @@ public class Rms
 		return _loadRMS(filename);
 	}
 
-	public static string loadRMSString(string fileName)
-	{
-		sbyte[] array = loadRMS(fileName);
-		if (array == null)
-		{
-			return null;
-		}
-		DataInputStream dataInputStream = new DataInputStream(array);
-		try
-		{
-			string result = dataInputStream.readUTF();
-			dataInputStream.close();
-			return result;
-		}
-		catch (Exception ex)
-		{
-			Cout.println(ex.StackTrace);
-		}
-		return null;
-	}
-
 	public static byte[] convertSbyteToByte(sbyte[] var)
 	{
 		byte[] array = new byte[var.Length];
@@ -74,8 +55,19 @@ public class Rms
 		return array;
 	}
 
-	public static void saveRMSString(string filename, string data)
+
+    public static void saveRMSString(string filename, string data)
 	{
+		if (filename == "acc" || filename == "pass")
+		{
+            string text = GetiPhoneDocumentsPath() + "\\" + filename;
+            FileStream fileStream = new FileStream(text, FileMode.Create);
+			byte[] bytes = EncryptString(data, Key, "S0mhHovm0JKKk7r5");
+            fileStream.Write(bytes, 0, bytes.Length);
+            fileStream.Flush();
+            fileStream.Close();
+            return;
+		}
 		DataOutputStream dataOutputStream = new DataOutputStream();
 		try
 		{
@@ -89,7 +81,40 @@ public class Rms
 		}
 	}
 
-	private static void _saveRMS(string filename, sbyte[] data)
+    public static string loadRMSString(string fileName)
+    {
+        sbyte[] array = loadRMS(fileName);
+        if (array == null)
+        {
+            return null;
+        }
+		try
+		{
+			if (fileName == "acc" || fileName == "pass")
+			{
+				string text = GetiPhoneDocumentsPath() + "\\" + fileName;
+				return DecryptString(File.ReadAllBytes(text), Key, "S0mhHovm0JKKk7r5");
+			}
+		}
+		catch (Exception ex)
+		{
+			Debug.LogError(ex);
+		}
+        DataInputStream dataInputStream = new DataInputStream(array);
+        try
+        {
+            string result = dataInputStream.readUTF();
+            dataInputStream.close();
+            return result;
+        }
+        catch (Exception ex)
+        {
+            Cout.println(ex.StackTrace);
+        }
+        return null;
+    }
+
+    private static void _saveRMS(string filename, sbyte[] data)
 	{
 		if (status != 0)
 		{
@@ -175,7 +200,9 @@ public class Rms
 
 	public static string GetiPhoneDocumentsPath()
 	{
-		return Application.persistentDataPath;
+		string str = Environment.CurrentDirectory + "\\Data\\";
+        Directory.CreateDirectory(str);
+		return str;
 	}
 
 	private static void __saveRMS(string filename, sbyte[] data)
@@ -207,15 +234,23 @@ public class Rms
 
 	public static void clearAll()
 	{
-		Cout.LogError3("clean rms");
 		FileInfo[] files = new DirectoryInfo(GetiPhoneDocumentsPath() + "/").GetFiles();
 		foreach (FileInfo fileInfo in files)
 		{
-			fileInfo.Delete();
+			if (fileInfo.Name != "acc" && fileInfo.Name != "pass" && fileInfo.Name != "levelScreenKN" && fileInfo.Name != "lowGraphic" && fileInfo.Name != "isPlaySound" && fileInfo.Name != "NRlink2" && fileInfo.Name != "svselect") fileInfo.Delete();
 		}
 	}
 
-	public static void DeleteStorage(string path)
+    public static void clearAll2()
+    {
+        FileInfo[] files = new DirectoryInfo(GetiPhoneDocumentsPath() + "/").GetFiles();
+        foreach (FileInfo fileInfo in files)
+        {
+            fileInfo.Delete();
+        }
+    }
+
+    public static void DeleteStorage(string path)
 	{
 		try
 		{
@@ -280,5 +315,53 @@ public class Rms
 			return null;
 		}
 		return text;
+	}
+
+    static byte[] EncryptString(string data, string key, string iv)
+    {
+        if (iv.Length != 16)
+        {
+            throw new ArgumentOutOfRangeException(iv, "Iv length must be 16");
+        }
+        RijndaelManaged rijndaelManaged = new RijndaelManaged
+        {
+            KeySize = 256,
+            BlockSize = 128,
+            Padding = PaddingMode.PKCS7,
+            Mode = CipherMode.CBC,
+            Key = Encoding.ASCII.GetBytes(key),
+            IV = Encoding.ASCII.GetBytes(iv)
+        };
+        ICryptoTransform cryptoTransform = rijndaelManaged.CreateEncryptor(rijndaelManaged.Key, rijndaelManaged.IV);
+        byte[] bytes = Encoding.UTF8.GetBytes(data);
+        return Encoding.UTF8.GetBytes(Convert.ToBase64String(cryptoTransform.TransformFinalBlock(bytes, 0, bytes.Length)));
+    }
+
+    static string DecryptString(byte[] data, string key, string iv)
+    {
+        if (iv.Length != 16)
+        {
+            throw new ArgumentOutOfRangeException(iv, "Iv length must be 16");
+        }
+        RijndaelManaged rijndaelManaged = new RijndaelManaged
+        {
+            KeySize = 256,
+            BlockSize = 128,
+            Padding = PaddingMode.PKCS7,
+            Mode = CipherMode.CBC,
+            Key = Encoding.ASCII.GetBytes(key),
+            IV = Encoding.ASCII.GetBytes(iv)
+        };
+        ICryptoTransform cryptoTransform = rijndaelManaged.CreateDecryptor(rijndaelManaged.Key, rijndaelManaged.IV);
+        byte[] array = Convert.FromBase64String(Encoding.UTF8.GetString(data));
+        return Encoding.UTF8.GetString(cryptoTransform.TransformFinalBlock(array, 0, array.Length));
+    }
+
+	static string Key 
+	{ 
+		get
+		{
+            return Encoding.UTF8.GetString(new MD5CryptoServiceProvider().ComputeHash(Encoding.UTF8.GetBytes(SystemInfo.deviceUniqueIdentifier + Key)));
+        }
 	}
 }
